Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-genai==1.14.0
pydantic==2.10.4
python-dotenv==1.0.1
